backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Financial Data APIs
yfinance==0.2.28
alpha-vantage==2.3.1

# Charts and Visualization
plotly==5.17.0
kaleido==0.2.1

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Excel Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Date and Time
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Code Quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# Utilities
python-slugify==8.0.1
pillow==10.1.0

# Development
watchdog==3.0.0